chore: refactor playground
